requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
aiogram==3.4.1

SQLAlchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2

pydantic==2.5.3
python-dotenv==1.0.1

httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.23.7

loguru==0.7.2

black==24.8.0
flake8==7.0.0
mypy==1.11.1
isort==5.13.2